+++ tests/test_communication.py
@@ -54,6 +54,29 @@ def test_connect_servo_ethernet_no_dictionary_error(read_config):
 
 
 @pytest.mark.smoke
+@pytest.mark.eoe
+@pytest.mark.parametrize(
+    "coco_dict_path",
+    [
+        True,
+        False,
+    ],
+)
+def test_connect_servo_comkit_no_dictionary_error(coco_dict_path, read_config):
+    mc = MotionController()
+    if coco_dict_path:
+        coco_dict_path = read_config["dictionary"]
+        moco_dict_path = "no_dictionary"
+    else:
+        coco_dict_path = "no_dictionary"
+        moco_dict_path = read_config["dictionary"]
+    with pytest.raises(FileNotFoundError):
+        mc.communication.connect_servo_comkit(
+            read_config["ip"], coco_dict_path, moco_dict_path, alias="eoe_test"
+        )
+
+
+@pytest.mark.smoke
 @pytest.mark.soem
 def test_get_ifname_from_interface_ip(mocker):
     ip = type("IP", (object,), {"ip": "192.168.2.1", "is_IPv4": True})